% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snow_variables.R
\name{cal_scd}
\alias{cal_scd}
\title{Internal Function to Calculate Snow Melting Day and Trend}
\usage{
cal_scd(d_all, extradf)
}
\arguments{
\item{d_all}{A data frame containing snow and no-snow values for a given cell.}

\item{extradf}{An optional data frame containing extra days for prediction.}
}
\value{
A data frame containing the calculated snow variables.
}
\description{
This function calculates the snow melting day (SCD) and trend for a given cell.
}
\keyword{internal}
