% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{calc_coverages}
\alias{calc_coverages}
\title{Internal Function to calculate cloud cover within the rasters}
\usage{
calc_coverages(image, image_dir)
}
\description{
Internal Function to calculate cloud cover within the rasters
}
\keyword{internal}
