% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{check_raster}
\alias{check_raster}
\title{Internal Function to check if raster files work properly}
\usage{
check_raster(image, image_dir)
}
\description{
Internal Function to check if raster files work properly
}
\keyword{internal}
