% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{extract_landsat}
\alias{extract_landsat}
\title{Extracts Landsat Imagery}
\usage{
extract_landsat(
  aoi,
  site_name,
  aoi_size = 5,
  start_date = "2013-01-01",
  end_date = Sys.Date(),
  months = 1:12,
  sats = "LC08",
  minclouds = 50,
  my_timeout = 400,
  info_timeout = 60,
  base_landsat_dir,
  workers,
  data_source = "rstac",
  force = FALSE
)
}
\arguments{
\item{aoi}{An \code{sf} object or a list containing the area of interest.}

\item{site_name}{A character string representing the name of the site.}

\item{aoi_size}{A numeric value representing the size of the AOI in kilometers (default is 5).}

\item{start_date}{A character string representing the start date for the imagery (default is "2013-01-01").}

\item{end_date}{A character string representing the end date for the imagery (default is the current date).}

\item{months}{A numeric vector representing the months to include (default is 1:12).}

\item{sats}{A character string representing the satellite to use (default is "LC08").}

\item{minclouds}{A numeric value representing the minimum cloud cover percentage (default is 50).}

\item{base_landsat_dir}{A character string representing the base directory for Landsat imagery.}

\item{workers}{A numeric value representing the number of workers for parallel processing.}

\item{data_source}{A character string representing the data source (default is "rstac").}

\item{force}{Logical, if TRUE deletes all previously downloaded imagery.}
}
\value{
A tibble containing the metadata of the downloaded Landsat scenes.
}
\description{
This function extracts Landsat imagery that fulfills the specified arguments.
It processes the area of interest (AOI), transforms coordinates, and downloads
the relevant Landsat scenes.
}
\examples{
\dontrun{
extract_landsat(aoi, site_name = "ExampleSite")
}
}
