% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{extract_landsat_stac}
\alias{extract_landsat_stac}
\title{Internal Function to Extract Landsat Imagery from STAC}
\usage{
extract_landsat_stac(
  aoi,
  epsg,
  excl_dates,
  site_name,
  sats = c("LT04", "LT05", "LE07", "LC08", "LC09"),
  start_date,
  end_date,
  months = 1:12,
  minclouds,
  area_landsat_dir,
  workers
)
}
\arguments{
\item{aoi}{An \code{sf} object representing the area of interest.}

\item{epsg}{A numeric value representing the EPSG code for the coordinate reference system.}

\item{excl_dates}{A tibble containing dates to exclude.}

\item{site_name}{A character string representing the name of the site.}

\item{sats}{A character vector representing the satellites to use (default is c("LT04","LT05","LE07","LC08","LC09")).}

\item{start_date}{A character string representing the start date for the imagery.}

\item{end_date}{A character string representing the end date for the imagery.}

\item{months}{A numeric vector representing the months to include.}

\item{minclouds}{A numeric value representing the minimum cloud cover percentage.}

\item{area_landsat_dir}{A character string representing the directory where Landsat imagery will be saved.}

\item{workers}{A numeric value representing the number of workers for parallel processing.}
}
\value{
A tibble containing the metadata of the downloaded Landsat scenes.
}
\description{
This function extracts Landsat imagery from the SpatioTemporal Asset Catalog (STAC)
based on the specified parameters. It is used internally by other functions.
}
\keyword{internal}
