% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sentinel2.R
\name{extract_sentinel2}
\alias{extract_sentinel2}
\title{Extracts Sentinel-2 Imagery}
\usage{
extract_sentinel2(
  aoi,
  site_name,
  aoi_size = 2,
  start_date = "2020-01-01",
  end_date = "2024-12-01",
  months = 1:12,
  sats = c("S2A", "S2B"),
  minclouds = 50,
  base_sentinel2_dir,
  workers = 1,
  data_source = "rstac",
  force = FALSE
)
}
\arguments{
\item{aoi}{An \code{sf} object or a list containing the area of interest.}

\item{site_name}{A character string representing the name of the site.}

\item{aoi_size}{A numeric value representing the size of the AOI in kilometers. Ignored if aoi is a POLYGON.}

\item{start_date}{A character string representing the start date for the imagery.}

\item{end_date}{A character string representing the end date for the imagery.}

\item{months}{A numeric vector representing the months to include (default is all months, so 1:12).}

\item{sats}{A character vector representing the satellite(s) to use (default is both, so c("S2A", "S2B")).}

\item{minclouds}{A numeric value representing the minimum cloud cover percentage (default is 50).}

\item{base_sentinel2_dir}{A character string representing the base directory for Sentinel-2 imagery.}

\item{workers}{A numeric value representing the number of workers for parallel processing.
Setting workers > 1 (default) enables parallel computing.}

\item{data_source}{A character string representing the data source (currently only option is "rstac").}

\item{force}{Logical, if TRUE deletes all previously downloaded imagery.}
}
\value{
A tibble containing the metadata of the downloaded Sentinel-2 scenes.
}
\description{
This function extracts Sentinel-2 L2A imagery that fulfills the specified arguments.
It processes the area of interest (AOI), transforms coordinates, and downloads
the relevant Sentinel-2 scenes from Microsoft Planetary Computer.
}
