% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{make_vsicurl_url}
\alias{make_vsicurl_url}
\title{Internal Function to Create a Base URL with Microsoft Planetary Computer}
\usage{
make_vsicurl_url(base_url)
}
\description{
Internal Function to Create a Base URL with Microsoft Planetary Computer
}
\keyword{internal}
