% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictors.R
\name{make_vsicurl_url_esa}
\alias{make_vsicurl_url_esa}
\title{Internal Function to Create a Base URL with Microsoft Planetary Computer}
\usage{
make_vsicurl_url_esa(base_url)
}
\description{
Internal Function to Create a Base URL with Microsoft Planetary Computer
}
\keyword{internal}
