% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictors.R
\name{mask_fill}
\alias{mask_fill}
\title{Internal Functions to extract pixel information from the Landsat Quality Assesment (QA) layer}
\usage{
mask_fill(x)
}
\description{
Internal Functions to extract pixel information from the Landsat Quality Assesment (QA) layer
}
\keyword{internal}
