% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snow_variables.R
\name{nosnow_values}
\alias{nosnow_values}
\title{Internal Function to Prepare No-Snow Values}
\usage{
nosnow_values(x)
}
\arguments{
\item{x}{A SpatRaster object.}
}
\value{
A data frame containing no-snow and snow values.
}
\description{
This function prepares no-snow values for further processing.
}
\keyword{internal}
