% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{process_feature}
\alias{process_feature}
\title{Internal Function to Extract Landsat Imagery from STAC}
\usage{
process_feature(ft, area_landsat_dir, aoi, tempdir)
}
\description{
Internal Function to Extract Landsat Imagery from STAC
}
\keyword{internal}
