% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sentinel2.R
\name{process_feature_s2}
\alias{process_feature_s2}
\title{Internal Function: Single Sentinel-2 Feature Extraction with Retry Logic}
\usage{
process_feature_s2(ft, area_sentinel2_dir, aoi, exte, reso)
}
\description{
Internal Function: Single Sentinel-2 Feature Extraction with Retry Logic
}
