% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{process_features_in_parallel}
\alias{process_features_in_parallel}
\title{Internal Function to Extract Landsat Imagery from STAC}
\usage{
process_features_in_parallel(it_obj, area_landsat_dir, aoi, workers)
}
\description{
Internal Function to Extract Landsat Imagery from STAC
}
\keyword{internal}
