% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sentinel2.R
\name{process_features_in_parallel_s2}
\alias{process_features_in_parallel_s2}
\title{Internal Function: Parallel Feature Processing (Sentinel-2) with Grid Alignment}
\usage{
process_features_in_parallel_s2(it_obj, area_sentinel2_dir, aoi, workers, epsg)
}
\description{
Internal Function: Parallel Feature Processing (Sentinel-2) with Grid Alignment
}
