% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snow_variables.R
\name{read_classifications}
\alias{read_classifications}
\title{Internal Function to Read Classifications}
\usage{
read_classifications(imagedf, basedir)
}
\description{
Internal Function to Read Classifications
}
\keyword{internal}
