% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_image_df.R
\name{search_image_df}
\alias{search_image_df}
\title{Extract image info}
\usage{
search_image_df(site_name, base_landsat_dir, workers = 1)
}
\arguments{
\item{site_name}{A character string representing the name of the site.}

\item{base_landsat_dir}{A character string representing the base directory for Landsat imagery.}

\item{workers}{A numeric value representing the number of workers for parallel processing.}

\item{data_source}{A character string representing the data source (default is "rstac").}
}
\value{
A tibble containing the metadata of the previously downloaded Landsat scenes.
}
\description{
This function extracts metadata from previously downloaded Landsat imagery.
}
\examples{
search_image_df(site_name = "ExampleSite")
}
