% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{snowman_cleanup}
\alias{snowman_cleanup}
\title{Quality Control: Reset Workers
Call this function after your loop rounds are completely finished.}
\usage{
snowman_cleanup()
}
\description{
Quality Control: Reset Workers
Call this function after your loop rounds are completely finished.
}
