% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{utm_zones}
\alias{utm_zones}
\title{UTM zones Geopackage file}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{ZONE}{The UTM zone identifier}
}
}
\source{
Created for demonstration purposes.
}
\description{
A Geopackage file with UTM zones as polygons.
}
\keyword{datasets}
