% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snow_variables.R
\name{cal_scd}
\alias{cal_scd}
\title{Internal Function to Calculate Snow Melting Day and Trend}
\usage{
cal_scd(d_all, extradf)
}
\description{
Internal Function to Calculate Snow Melting Day and Trend
}
\keyword{internal}
