% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predictors.R
\name{calc_predictors}
\alias{calc_predictors}
\title{Calculate Predictors for Landsat Imagery}
\usage{
calc_predictors(image_df, site_name, base_landsat_dir, data_source = "rstac")
}
\arguments{
\item{image_df}{A tibble, the output from the \code{extract_landsat} or \code{search_image_df} functions.}

\item{site_name}{A character string representing the name of the site.}

\item{base_landsat_dir}{A character string representing the base directory for Landsat imagery.}

\item{data_source}{A character string representing the data source (default is "rstac").}
}
\value{
Raster files written to disk.
}
\description{
This function calculates various predictors from Landsat imagery and other data sources.
It processes the input image data frame, downloads necessary data, and computes several
predictors such as DEM variables, median Landsat images, and median indices.
}
\examples{
\dontrun{
calc_predictors(image_df, site_name = "ExampleSite", base_landsat_dir = "path/to/landsat")
}
}
