% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snow_variables.R
\name{calc_snow_variables}
\alias{calc_snow_variables}
\title{Calculate Snow Variables}
\usage{
calc_snow_variables(image_df, site_name, base_landsat_dir, workers)
}
\arguments{
\item{image_df}{A tibble, the output from the \code{extract_landsat} or \code{search_image_df} functions.}

\item{site_name}{A character string representing the name of the site.}

\item{base_landsat_dir}{A character string representing the base directory for Landsat imagery.}

\item{workers}{A numeric value representing the number of workers for parallel processing.}
}
\value{
A SpatRaster object containing the snow variables.
}
\description{
This function calculates snow variables from classified Landsat imagery. It processes the input image data frame,
calculates snow proportions, and generates snow melting day (SCD) and trend rasters.
}
\examples{
calc_snow_variables(image_df, site_name = "ExampleSite", base_landsat_dir = "path/to/landsat")
}
