% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{check_raster}
\alias{check_raster}
\title{Internal Function to check if raster files work properly}
\usage{
check_raster(image, image_dir)
}
\arguments{
\item{image}{A character string representing the file name.}

\item{image_dir}{A character string representing directory for the files.}
}
\value{
A character vector representing the raster files that did not work.
}
\description{
Internal Function to check if raster files work properly
}
\keyword{internal}
