% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_landsat.R
\name{classify_landsat}
\alias{classify_landsat}
\title{Classify Landsat Imagery}
\usage{
classify_landsat(
  image_df,
  site_name,
  base_landsat_dir,
  model_dir,
  workers,
  force = FALSE
)
}
\arguments{
\item{image_df}{A tibble, the output from the \code{extract_landsat} or \code{search_image_df} functions.}

\item{site_name}{A character string representing the name of the site.}

\item{base_landsat_dir}{A character string representing the base directory for Landsat imagery.}

\item{model_dir}{A character string representing the path to the directory containing the pretrained models.}

\item{workers}{A numeric value representing the number of workers for parallel processing.}

\item{force}{Logical, if TRUE classification will be run also for images with previously classified.}
}
\value{
A list of file paths to the classified rasters.
}
\description{
This function classifies Landsat imagery using pretrained models. It processes the input image data frame,
applies the models, and generates classification rasters.
}
\examples{
classify_landsat(image_df, site_name = "ExampleSite", base_landsat_dir = "path/to/landsat", model_dir = "path/to/models")
}
