% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_landsat.R
\name{classifying_function}
\alias{classifying_function}
\title{Internal Function to Classify a Single Landsat Image}
\usage{
classifying_function(
  imageid,
  image_df,
  predictor_dir,
  class_landsat_dir,
  base_landsat_dir,
  site_name,
  force = FALSE,
  mod8,
  mod7,
  mod5,
  slope,
  esalc,
  esaw,
  dem
)
}
\description{
Internal Function to Classify a Single Landsat Image
}
\keyword{internal}
