% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{extract_landsat_stac}
\alias{extract_landsat_stac}
\title{Internal Function to Extract Landsat Imagery from STAC}
\usage{
extract_landsat_stac(
  aoi,
  epsg,
  excl_dates,
  site_name,
  sats = c("LT04", "LT05", "LE07", "LC08", "LC09"),
  start_date,
  end_date,
  months = 1:12,
  minclouds,
  area_landsat_dir,
  workers
)
}
\description{
Internal Function to Extract Landsat Imagery from STAC
}
\keyword{internal}
