% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{make_vsicurl_url}
\alias{make_vsicurl_url}
\title{Internal Function to Create a Base URL with Microsoft Planetary Computer}
\usage{
make_vsicurl_url(base_url)
}
\arguments{
\item{base_url}{A character string representing the STAC URL for the image.}
}
\value{
A character string representing the base URL to extract the targeted image.
}
\description{
This function creates a base URL for extracting targeted images from the Microsoft Planetary Computer.
It is used internally by other functions.
}
\keyword{internal}
