% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snow_variables.R
\name{nosnow_values}
\alias{nosnow_values}
\title{Internal Function to Prepare No-Snow Values}
\usage{
nosnow_values(x)
}
\description{
Internal Function to Prepare No-Snow Values
}
\keyword{internal}
