% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landsat.R
\name{process_feature}
\alias{process_feature}
\title{Internal Function to Extract Landsat Imagery from STAC}
\usage{
process_feature(ft, area_landsat_dir, aoi, tempdir)
}
\arguments{
\item{ft}{A STAC feature queried from MPC.}

\item{area_landsat_dir}{A character string representing the directory where Landsat imagery will be saved.}

\item{aoi}{An \code{sf} object representing the area of interest.}

\item{tempdir}{A character string representing the temporary directory where Landsat imagery will be saved.}
}
\value{
FALSE or TRUE depending on the success of the downloading.
}
\description{
Internal Function to Extract Landsat Imagery from STAC
}
\keyword{internal}
