% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snow_variables.R
\name{read_classifications}
\alias{read_classifications}
\title{Internal Function to Read Classifications}
\usage{
read_classifications(imagedf, basedir)
}
\arguments{
\item{imagedf}{A tibble containing the metadata of the classified images.}

\item{basedir}{A character string representing the base directory for the classified images.}
}
\value{
A list of SpatRaster objects.
}
\description{
This function reads classified rasters and prepares them for further processing.
}
\keyword{internal}
