% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utm_zones}
\alias{utm_zones}
\title{UTM Zones}
\format{
A data frame with X rows and Y variables:
\describe{
\item{zone}{Zone number}
\item{hemisphere}{"N" or "S"}
\item{epsg}{EPSG code for the zone}
...
}
}
\source{
Source of the data, if applicable
}
\usage{
utm_zones
}
\description{
A dataset containing the Universal Transverse Mercator (UTM) zone information.
}
\keyword{datasets}
